Moves ArtboardNavigation into Artboard Dir
