+++ lib/pages/Artboard/Artboard.js
@@ -1,10 +1,10 @@
 import styled from "styled-components";
-import { Navigation } from "../components/Layout/Navigation/Navigation";
-import { Main } from "../components/Layout/Main";
+import { Navigation } from "../../components/Layout/Navigation/Navigation";
+import { Main } from "../../components/Layout/Main";
 import { useRouter } from "next/router";
 import { useQuery } from "@apollo/client";
-import { GET_DOCUMENT } from "../gql/queries/getDocumentQuery";
-import { ArtboardNavigation } from "../components/ArtboardNavigation/ArtboardNavigation";
+import { GET_DOCUMENT } from "../../gql/queries/getDocumentQuery";
+import { ArtboardNavigation } from "./ArtboardNavigation/ArtboardNavigation";
 
 const ArtboardTitle = styled(Navigation.Title)`
   flex: 0 1 auto;
