+++ lib/pages/Artboard/ArtboardNavigation/ArtboardNavigation.js
@@ -1,6 +1,6 @@
 import Link from "next/link";
 import styled from "styled-components";
-import { Navigation } from "../Layout/Navigation/Navigation";
+import { Navigation } from "../../../components/Layout/Navigation/Navigation";
 import { useRouter } from "next/router";
 
 const ArtboardCountText = styled.p`
