+++ pages/documents/[documentShortId]/artboards/[artboardIndex].js
@@ -1,3 +1,3 @@
-import { Artboard } from "../../../../lib/pages/Artboard";
+import { Artboard } from "../../../../lib/pages/Artboard/Artboard";
 
 export default Artboard;
